requirements.txt

scikit-surprise
scikit-learn
pandas
 
